Added starship

+++ scripts/install_zsh.sh
@@ -2,9 +2,15 @@
 
 cd "$(dirname "$0")"
 
+curl -sS https://starship.rs/install.sh | sh
+
 sudo apt install zsh -y
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 chsh -s "/bin/zsh" "$USER"
-rm "$HOME/.zshrc"
+
+if [ -f "$HOME/.zshrc"]; then
+  rm "$HOME/.zshrc"
+fi
+
 cp "../.zshrc" "$HOME/.zshrc"
 zsh
